+++ config_template.py
@@ -21,6 +21,7 @@ HIVE_HOST = 'localhost'
 HIVE_PORT = 10000
 HIVE_DATABASE = 'default'
 HIVE_METASTORE_URI = 'postgresql://user:password@host:port/database'
+HIVE_YARN_QUEUE = 'default'
 
 # CosmoHub settings
 SQLALCHEMY_DATABASE_URI = 'postgresql://user:password@host:port/database'
@@ -28,11 +29,19 @@ SQLALCHEMY_TRACK_MODIFICATIONS = False
 DOWNLOADS_BASE_DIR = ''
 RESULTS_BASE_DIR = 'cosmohub_results' # Relative to ~
 
-# 64-byte (128 hex-chars) secret key for signing tokens and passwords
+# 64-byte (128 hex-chars) secret key for signing tokens and cookies
+# Change this to invalidate all sessions and tokens
 SECRET_KEY = binascii.unhexlify(
     '0000000000000000000000000000000000000000000000000000000000000000'
     '0000000000000000000000000000000000000000000000000000000000000000'
 )
+# Token expiration in seconds
+TOKEN_EXPIRES_IN = {
+    'email_confirm'  : float('+inf'),
+    'password_reset' : 2*24*60*60,
+    'download'       : float('+inf'),
+}
+TOKEN_EXPIRES_IN_DEFAULT = 1*60*60
 
 # Password settings
 PASSLIB_CONTEXT = {
@@ -40,9 +49,9 @@ PASSLIB_CONTEXT = {
         'pbkdf2_sha512'
     ],
 
-    'pbkdf2_sha512__min_rounds'     :   5000,
-    'pbkdf2_sha512__default_rounds' :  50000,
-    'pbkdf2_sha512__max_rounds'     : 500000,
+    'pbkdf2_sha512__min_rounds'     :   5000, # Takes 0.01 s
+    'pbkdf2_sha512__default_rounds' :  50000, # Takes 0.1 s
+    'pbkdf2_sha512__max_rounds'     : 500000, # Takes 1 s
     'pbkdf2_sha512__vary_rounds'    :    0.1,
     'pbkdf2_sha512__salt_size'      :     64,
 }
@@ -50,11 +59,9 @@ PASSLIB_CONTEXT = {
 # WebHCat settings
 WEBHCAT_BASE_URL = 'http://localhost:50111/templeton/v1/'
 WEBHCAT_SCRIPT_COMMON = textwrap.dedent("""\
-    # Workaround for HIVE-11607. Fixed in 1.3.0 and 2.0.0
     SET hive.exec.copyfile.maxsize=1073741824;
     """
 )
-WEBHCAT_SCRIPT_ROW_FORMAT = [ 'csv.bz2' ]
 WEBHCAT_SCRIPT_TEMPLATE = textwrap.dedent("""\
     {common_config}
     {compression_config}
@@ -65,3 +72,20 @@ WEBHCAT_SCRIPT_TEMPLATE = textwrap.dedent("""\
     ;
     """
 )
+
+MAIL_SERVER = 'localhost'
+MAIL_PORT = 25
+MAIL_USE_TLS = False
+MAIL_USE_SSL = False
+MAIL_USERNAME = None
+MAIL_PASSWORD = None
+MAIL_DEFAULT_SENDER = 'CosmoHub <cosmohub@pic.es>'
+
+RECAPTCHA_ENABLED = True
+RECAPTCHA_SITE_KEY = ''
+RECAPTCHA_SECRET_KEY = ''
+
+if DEBUG:
+    GA_TRACKING_ID = ''
+else:
+    GA_TRACKING_ID = ''