+++ cosmohub/api/__init__.py
@@ -5,12 +5,15 @@ patch_psycopg()
 
 import logging
 
-from flask import Flask, Blueprint, jsonify
+from flask import g, Flask, Blueprint, jsonify
 from flask_logconfig import LogConfig
+from flask_mail import Mail
+from flask_recaptcha import ReCaptcha
 from flask_restful import Api
 from flask_sqlalchemy import SQLAlchemy, Model as FlaskModel
 from flask_sockets import Sockets
 from flask_cors import CORS
+from itsdangerous import TimedJSONWebSignatureSerializer
 from pkg_resources import iter_entry_points # @UnresolvedImport
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.ext import declarative
@@ -24,11 +27,17 @@ log = logging.getLogger(__name__)
 app = Flask(__name__)
 app.config.from_object('config')
 
-# Enable CORS
-CORS(app)
-
 # Configure logging
 logconfig = LogConfig(app)
+
+# Enable CORS
+CORS(app, expose_headers=['X-Token'])
+
+# Configure mail service
+mail = Mail(app)
+
+# Configure ReCaptcha service
+recaptcha = ReCaptcha(app)
 
 # Configure SQLAlchemy extension
 metadata = schema.MetaData()
@@ -53,6 +62,29 @@ app.formats = {
     entry.name : entry.load()
     for entry in iter_entry_points(group='cosmohub_format')
 }
+
+# Set up token signer
+app.jwt = TimedJSONWebSignatureSerializer(app.config['SECRET_KEY'])
+
+@app.before_request
+def _clear_session():
+    g.session = {
+        'user' : None,
+        'privilege' : None,
+        'token' : None,
+    }
+
+# Add/refresh token to every authenticated request 
+from .security import Token, Privilege
+# FIXME: refactor authentication code to remove cicle import on db
+@app.after_request
+def _refresh_token(response):
+    if g.session['privilege'] and Privilege(['user']).can(g.session['privilege']):
+        response.headers['X-Token'] = Token(g.session['user'], g.session['privilege']).dump()
+    else:
+        response.headers['X-Token'] = g.session['token']
+    
+    return response
 
 # Configure REST API Blueprint
 mod_rest = Blueprint('rest', __name__, url_prefix='/rest')