+++ cosmohub/api/fields.py
@@ -1,15 +1,17 @@
 from flask_restful import fields
 
-USER = {
-    'id'         : fields.Integer,
-    'name'       : fields.String,
-    'email'      : fields.String,
-    'is_enabled' : fields.Boolean,
-    'is_admin'   : fields.Boolean,
-    'groups'     : fields.List(fields.String(attribute='name')),
+User = {
+    'id'                 : fields.Integer,
+    'name'               : fields.String,
+    'email'              : fields.String,
+    'is_admin'           : fields.Boolean,
+    'groups'             : fields.List(fields.String(attribute='name')),
+    'ts_created'         : fields.DateTime('iso8601'),
+    'ts_email_confirmed' : fields.DateTime('iso8601'),
+    'ts_last_login'      : fields.DateTime('iso8601'),
 }
 
-CATALOGS = {
+CatalogCollection = {
     'id'           : fields.Integer,
     'name'         : fields.String,
     'version'      : fields.String,
@@ -20,49 +22,48 @@ CATALOGS = {
     'ts_uploaded'  : fields.DateTime('iso8601'),
 }
 
-DATASET = {
-    'id'          : fields.Integer,
-    'name'        : fields.String,
-    'version'     : fields.String,
-    'description' : fields.String,
-    'rows'        : fields.Integer,
-    'recipe'      : fields.Raw,
-    'path_readme' : fields.String,
-    'ts_defined'  : fields.DateTime('iso8601'),
+Dataset = {
+    'id'              : fields.Integer,
+    'name'            : fields.String,
+    'version'         : fields.String,
+    'description'     : fields.String,
+    'rows'            : fields.Integer,
+    'recipe'          : fields.Raw,
+    'ts_defined'      : fields.DateTime('iso8601'),
 }
 
-FILE = {
-    'id'            : fields.Integer,
-    'name'          : fields.String,
-    'version'       : fields.String,
-    'description'   : fields.String,
-    'size'          : fields.Integer,
-    'path_readme'   : fields.String,
-    'path_contents' : fields.String,
-    'ts_uploaded'   : fields.DateTime('iso8601'),
+File = {
+    'id'                : fields.Integer,
+    'name'              : fields.String,
+    'version'           : fields.String,
+    'description'       : fields.String,
+    'size'              : fields.Integer,
+    'ts_uploaded'       : fields.DateTime('iso8601'),
 }
 
-CATALOG = CATALOGS.copy()
-CATALOG.update({
-    'relation' : fields.String,
+Catalog = CatalogCollection.copy()
+Catalog.update({
     'summary'  : fields.String,
-    'datasets' : fields.List(fields.Nested(DATASET)),
-    'files'    : fields.List(fields.Nested(FILE)),
+    'relation' : fields.String,
+    'rows'     : fields.Integer,
+    'datasets' : fields.List(fields.Nested(Dataset)),
+    'files'    : fields.List(fields.Nested(File)),
 })
 
-COLUMN = {
-    'name' : fields.String,
-    'type' : fields.String,
+Query = {
+    'id'               : fields.Integer,
+    'sql'              : fields.String,
+    'format'           : fields.String,
+    'status'           : fields.String,
+    'job_id'           : fields.String,
+    'size'             : fields.Integer,
+    'ts_submitted'     : fields.DateTime('iso8601'),
+    'ts_started'       : fields.DateTime('iso8601'),
+    'ts_finished'      : fields.DateTime('iso8601'),
 }
 
-QUERY = {
-    'id'           : fields.Integer,
-    'sql'          : fields.String,
-    'format'       : fields.String,
-    'status'       : fields.String,
-    'job_id'       : fields.String,
-    'size'         : fields.Integer,
-    'ts_submitted' : fields.DateTime('iso8601'),
-    'ts_started'   : fields.DateTime('iso8601'),
-    'ts_finished'  : fields.DateTime('iso8601'),
+UserToken = {
+    'id'         : fields.Integer,
+    'name'       : fields.String,
+    'email'      : fields.String,
 }