+++ CHANGELOG.md
@@ -6,6 +6,32 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased][unreleased]
 
 
+## [1.1.0] - 2016-09-05
+### Added
+- Return total number of rows of a catalog. (Pau Tallada)
+- Token authentication and authorization. (Pau Tallada)
+- Send interactive queries to custom `YARN` queue. (Pau Tallada)
+- Transaction emails on register and password reset. (Pau Tallada)
+- Generate download links with tokens for files and datasets. (Pau Tallada)
+- Initial transactional mail templates. (Pau Tallada)
+- Capcha validation on password reset and user registration. (Pau Tallada)
+- Allow cancelling of a running query. (Pau Tallada)
+
+### Changed
+- Generate download links with tokens for query results. (Pau Tallada)
+- Add `ts_last_updated` with a trigger to ACL table. (Pau Tallada)
+- Retrieve redirect URL from frontend in user registration. (Pau Tallada)
+- Track new and completed queries on `queries` websocket. (Pau Tallada)
+- Refactor all authentication code. (Pau Tallada)
+
+### Fixed
+- Return partitioned table statistics. (Pau Tallada)
+- Store query start and finish time in UTC. (Pau Tallada)
+- Return partitions as virtual columns. (Pau Tallada)
+- Handle only failed jobs in hive progress. (Pau Tallada)
+- Authenticate web sockets connections. (Pau Tallada)
+
+
 ## [1.0.0] - 2016-08-19
 ### Added
 - Implement a new data model. (Pau Tallada)